switch to BigDecimal

+++ src/main/java/pl/andrzejjozefow/orderbook/Deal.java
@@ -1,12 +1,13 @@
 package pl.andrzejjozefow.orderbook;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class Deal {
     final User buyer;
     final User seller;
-    final Double price;
+    final BigDecimal price;
     final Integer quantity;
 
     @Override
